Fix findbugs warning in IndexJob

+++ molgenis-data-elasticsearch/src/main/java/org/molgenis/data/elasticsearch/index/job/IndexJob.java
@@ -11,7 +11,6 @@ import org.molgenis.data.jobs.Job;
 import org.molgenis.data.jobs.Progress;
 import org.molgenis.data.meta.model.EntityType;
 import org.molgenis.data.meta.model.EntityTypeFactory;
-import org.molgenis.data.meta.model.EntityTypeMetadata;
 import org.molgenis.data.support.QueryImpl;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -127,7 +126,8 @@ class IndexJob extends Job
 			{
 				if (indexAction.getEntityId() != null)
 				{
-					progress.progress(progressCount, format("Indexing {0}.{1}", entityType.getFullyQualifiedName(), indexAction.getEntityId()));
+					progress.progress(progressCount,
+							format("Indexing {0}.{1}", entityType.getFullyQualifiedName(), indexAction.getEntityId()));
 					rebuildIndexOneEntity(entityTypeId, indexAction.getEntityId());
 				}
 				else
@@ -142,14 +142,17 @@ class IndexJob extends Job
 				entityType = getEntityType(indexAction);
 				if (searchService.hasMapping(entityType))
 				{
-					progress.progress(progressCount, format("Dropping entityType with name: {0} and ID: {1}", entityType.getName(), entityType.getId()));
+					progress.progress(progressCount,
+							format("Dropping entityType with name: {0} and ID: {1}", entityType.getName(),
+									entityType.getId()));
 					searchService.delete(entityType);
 				}
 				else
 				{
 					// Index Job is finished, here we concluded that we don't have enough info to continue the index job
 					progress.progress(progressCount,
-							format("Skip index entity {0}.{1}", entityType.getFullyQualifiedName(), indexAction.getEntityId()));
+							format("Skip index entity {0}.{1}", entityType.getFullyQualifiedName(),
+									indexAction.getEntityId()));
 				}
 			}
 			updateIndexActionStatus(indexAction, IndexActionMetaData.IndexStatus.FINISHED);
@@ -178,7 +181,7 @@ class IndexJob extends Job
 	/**
 	 * Indexes one single entity instance.
 	 *
-	 * @param entityTypeId  the id of the entity's repository
+	 * @param entityTypeId the id of the entity's repository
 	 * @param untypedEntityId the identifier of the entity to update
 	 */
 	private void rebuildIndexOneEntity(String entityTypeId, String untypedEntityId)
@@ -187,43 +190,48 @@ class IndexJob extends Job
 
 		// convert entity id string to typed entity id
 		EntityType entityType = dataService.getEntityTypeById(entityTypeId);
-		Object entityId =
-				entityType != null ? getTypedValue(untypedEntityId, entityType.getIdAttribute()) : untypedEntityId;
-
-		String entityFullName = entityType.getFullyQualifiedName();
-
-		Entity actualEntity = dataService.findOneById(entityFullName, entityId);
-
-		if (null == actualEntity)
+		if (null != entityType)
 		{
-			// Delete
-			LOG.debug("Index delete [{}].[{}].", entityFullName, entityId);
-			searchService.deleteById(entityId.toString(), entityType);
-			return;
-		}
+			Object entityId = getTypedValue(untypedEntityId, entityType.getIdAttribute());
+			String entityFullName = entityType.getFullyQualifiedName();
 
-		boolean indexEntityExists = searchService.hasMapping(entityType);
-		if (!indexEntityExists)
-		{
-			LOG.debug("Create mapping of repository [{}] because it was not exist yet", entityFullName);
-			searchService.createMappings(entityType);
-		}
+			Entity actualEntity = dataService.findOneById(entityFullName, entityId);
 
-		Query<Entity> q = new QueryImpl<>();
-		q.eq(entityType.getIdAttribute().getName(), entityId);
-		Entity indexEntity = searchService.findOne(q, entityType);
+			if (null == actualEntity)
+			{
+				// Delete
+				LOG.debug("Index delete [{}].[{}].", entityFullName, entityId);
+				searchService.deleteById(entityId.toString(), entityType);
+				return;
+			}
 
-		if (null != indexEntity)
-		{
-			// update
-			LOG.debug("Index update [{}].[{}].", entityFullName, entityId);
-			searchService.index(actualEntity, actualEntity.getEntityType(), IndexingMode.UPDATE);
+			boolean indexEntityExists = searchService.hasMapping(entityType);
+			if (!indexEntityExists)
+			{
+				LOG.debug("Create mapping of repository [{}] because it was not exist yet", entityTypeId);
+				searchService.createMappings(entityType);
+			}
+
+			Query<Entity> q = new QueryImpl<>();
+			q.eq(entityType.getIdAttribute().getName(), entityId);
+			Entity indexEntity = searchService.findOne(q, entityType);
+
+			if (null != indexEntity)
+			{
+				// update
+				LOG.debug("Index update [{}].[{}].", entityTypeId, entityId);
+				searchService.index(actualEntity, actualEntity.getEntityType(), IndexingMode.UPDATE);
+			}
+			else
+			{
+				// Add
+				LOG.debug("Index add [{}].[{}].", entityTypeId, entityId);
+				searchService.index(actualEntity, actualEntity.getEntityType(), IndexingMode.ADD);
+			}
 		}
 		else
 		{
-			// Add
-			LOG.debug("Index add [{}].[{}].", entityFullName, entityId);
-			searchService.index(actualEntity, actualEntity.getEntityType(), IndexingMode.ADD);
+			throw new MolgenisDataException("Unknown EntityType for entityTypeId: " + entityTypeId);
 		}
 	}
